backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
argon2-cffi==23.1.0
bcrypt==4.1.2
blake3==1.0.8
